backend/app/schemas.py
from pydantic import BaseModel, Field


class CardData(BaseModel):
    id: str = Field(min_length=1)
    title: str = Field(min_length=1)
    details: str = ""


class ColumnData(BaseModel):
    id: str = Field(min_length=1)
    title: str = Field(min_length=1)
    cardIds: list[str]


class BoardData(BaseModel):
    columns: list[ColumnData]
    cards: dict[str, CardData]


class AIConnectivityResponse(BaseModel):
    model: str
    prompt: str
    response: str
