backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "Backend service for PM MVP"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.121.0",
  "httpx>=0.28.1",
  "python-dotenv>=1.2.2",
  "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
